% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rosabox.R
\name{rosabox}
\alias{rosabox}
\title{rosabox}
\usage{
rosabox(df, var, ...)
}
\arguments{
\item{df}{a data frame}

\item{var}{a column vector from the data frame (numerical)}

\item{...}{passes extra arguments to the function}
}
\value{
a boxplot and histogram, plus a list of data from the `geom_boxplot` function output
}
\description{
Produces a boxplot and histogram of the df using `ggplot2`
}
\examples{
\dontrun{rosabox(ddt, LENGTH)}
}
